# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is trading matching service.",
        "title": "Trading Matching Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "paths": {
        "/orders": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "PlaceOrder",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.placeOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.placeOrderResponse"
                        }
                    }
                }
            }
        },
        "/orders/{oid}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "CancelOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "oid",
                        "name": "oid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GeneralResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.GeneralResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.placeOrderRequest": {
            "type": "object",
            "properties": {
                "order_kind": {
                    "description": "OrderKind:\n* 1 - buy order.\n* 2 - sell order.",
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "price_type": {
                    "description": "PriceType:\n* 1 - market price.\n* 2 - limit price.",
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "api.placeOrderResponse": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                }
            }
        }
    }
}